add all sides
RedLeft and BlueLeft (redRight and blueRight as well) should be interchangable, see parking positions and setup for y. Before building this, look @ step 7 on the EasyOpenCV and chage file on REV hub.

Red Left Side might have  some worng stuff where the motors r running with the sleeps and which direction the robot goes in omniDrive.